requirements.txt
# pip install -r requirements.txt
selenium
webdriver-manager
pyyaml